Replace the horrible awk hack in Makefile with sed
I seem to have forgotten that any character can be used as a delimiter
for sed so the text replacement to dynamically generate maotd's fortune
argument was previously done with a painfully hacky awk command and a
temporary file to get around make's interpretation of backslashes as
line continuation preventing escaping of slashes. Replaced that nonsense
with an actually reasonable sed command.